rename everywhere from filter to decode
